feat (board): Refactor task ids. Checks existing ids on load, and increments nextId on the current board past the maxId of the loaded tasks

+++ src/main/java/com/charlesfitzgerald/kanban/Task.java
@@ -6,7 +6,7 @@ public class Task {
     private String description;
     private int priority;
 
-    public Task(int id, String title, String description, int priority) {
+    public Task(long id, String title, String description, int priority) {
         this.id = id;
         this.title = title;
         this.description = description;
